move the uoms to arptypes
